Set up schedule API with mock data
This is so we can get the mobile guys going on their development, before
the data has actually been set in stone.

I also added middleware that logs every http request made to the server.

I added the 'shortid' package to the project, but then I realized I don't
need it right away. I'll keep it for later, when we do have to start
generating IDs for things.